Move EpisodesCacheTest to episode module and add database test dependency.
